adds the sidebar for the event pages as well
the lack of a mobile sidebar was causing the inability to buy tickets
from the event pages, since there wasn’t a good conversation upfront
about how the site structure should really get used.

IE, the pages are duplicated, but the shows should really be for each
event, not each artist for each event.